turned int to double

+++ app/src/main/java/com/example/weatherapp/model/Minutely.kt
@@ -4,6 +4,6 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class Minutely(
-    val dt: Int?,
-    val precipitation: Int?
+    val dt: Double?,
+    val precipitation: Double?
 )
